+++ src/features/vote/styled/UpcomingRequests.styled.tsx
@@ -1,0 +1,8 @@
+/** @jsxImportSource @emotion/react */
+import tw, { styled } from "twin.macro"; // eslint-disable-line
+import RequestsWrapper from "./RequestsWrapper.styled";
+
+export const Wrapper = styled(RequestsWrapper)`
+  &.UpcomingRequests {
+  }
+`;
